increase precision

+++ pydigree/stats/mixedmodel/likelihood.py
@@ -130,7 +130,7 @@ def reml_hessian(y, X, V, ranefs, P=None, Vinv=None):
             mat[i, j] = element
             mat[j, i] = element
 
-    return np.array(mat)
+    return np.array(mat, dtype=np.float128)
 
 
 def reml_observed_information_matrix(y, X, V, ranefs, P=None, Vinv=None):
@@ -147,7 +147,7 @@ def reml_fisher_information_matrix(y, X, V, ranefs, P=None, Vinv=None):
     if P is None:
         P = makeP(X, Vinv)
 
-    mat = np.zeros((len(ranefs), len(ranefs)))
+    mat = np.zeros((len(ranefs), len(ranefs)), dtype=np.float128)
 
     for i, ranef_a in enumerate(ranefs):
         dV_dsigma_a = ranef_a.V_i
@@ -174,7 +174,7 @@ def reml_average_information_matrix(y, X, V, ranefs, P=None, Vinv=None):
         Vinv = makeVinv(V)
     if P is None:
         P = makeP(X, Vinv)
-    mat = np.zeros((len(ranefs), len(ranefs)))
+    mat = np.zeros((len(ranefs), len(ranefs)), dtype=np.float128)
     for i, ranef_a in enumerate(ranefs):
         dV_dsigma_a = ranef_a.V_i
